# Sort legal moves by heuristic scores