--- a/cpp/src/game/include/game/Strategy.hpp
+++ b/cpp/src/game/include/game/Strategy.hpp
@@ -19,15 +19,15 @@ namespace Alphalcazar::Game {
 		virtual ~Strategy() = default;
 
 		/*!
-		 * \brief Returns the placement move index to execute given the state of the game 
+		 * \brief Returns the placement move to execute given the state of the game 
 		          and the player who executes the strategy.
 	     *
 		 * \param playerId The ID of the player who's turn it is to play.
 		 * \param legalMoves The list of available legal moves the player currently has available.
 		 * \param game The game object. Can be used to factor in the state of the game and the board in order to chose a move.
 		 * 
-		 * \returns The index of the move (in the legalMoves parameter) of the move to be executed.
+		 * \returns The move to be executed.
 		 */
-		virtual PlacementMoveIndex Execute(PlayerId playerId, const std::vector<PlacementMove>& legalMoves, const Game& game) = 0;
+		virtual PlacementMove Execute(PlayerId playerId, const std::vector<PlacementMove>& legalMoves, const Game& game) = 0;
 	};
 }