--- a/cpp/src/game/src/game/Game.cpp
+++ b/cpp/src/game/src/game/Game.cpp
@@ -71,14 +71,8 @@ namespace Alphalcazar::Game {
 		auto legalMoves = GetLegalMoves(playerId);
 		// If a player has no available legal moves, their turn is skipped
 		if (legalMoves.size() > 0) {
-			PlacementMoveIndex placementMoveIndex = strategy.Execute(playerId, legalMoves, *this);
-			if (placementMoveIndex >= 0 && placementMoveIndex < legalMoves.size()) {
-				auto& placementMove = legalMoves[placementMoveIndex];
-				ExecutePlacementMove(playerId, placementMove);
-			} else {
-				Utils::LogError("Invalid legal move index ({}) returned by player strategy", placementMoveIndex);
-			}
-
+			auto placementMove = strategy.Execute(playerId, legalMoves, *this);
+			ExecutePlacementMove(playerId, placementMove);
 		}
 	}