--- a/cpp/src/strategies/minmax/src/minmax/BoardEvaluation.cpp
+++ b/cpp/src/strategies/minmax/src/minmax/BoardEvaluation.cpp
@@ -73,7 +73,6 @@ namespace Alphalcazar::Strategy::MinMax {
 					totalScore -= pieceScore;
 				}
 			}
-
 		}
 		return totalScore;
 	}