--- a/cpp/src/strategies/minmax/tests/LegalMovements.cpp
+++ b/cpp/src/strategies/minmax/tests/LegalMovements.cpp
@@ -19,7 +19,7 @@ namespace Alphalcazar::Strategy::MinMax {
 		Game::Game game {};
 
 		// An empty board should have both X and Y symmetry
-		auto [xSymmetryEmpty, ySymmetryEmpty] = GetBoardSymmetries(game);
+		auto [xSymmetryEmpty, ySymmetryEmpty] = GetBoardSymmetries(game.GetBoard());
 		EXPECT_TRUE(xSymmetryEmpty);
 		EXPECT_TRUE(ySymmetryEmpty);
 
@@ -27,7 +27,7 @@ namespace Alphalcazar::Strategy::MinMax {
 		game.GetBoard().PlacePiece({ 2, 2 }, pieceOne, Game::Direction::NORTH);
 
 		// A piece in the center pointing north should create Y symmetry
-		auto [xSymmetryCenterPiece, ySymmetryCenterPiece] = GetBoardSymmetries(game);
+		auto [xSymmetryCenterPiece, ySymmetryCenterPiece] = GetBoardSymmetries(game.GetBoard());
 		EXPECT_FALSE(xSymmetryCenterPiece);
 		EXPECT_TRUE(ySymmetryCenterPiece);
 
@@ -35,7 +35,7 @@ namespace Alphalcazar::Strategy::MinMax {
 		game.GetBoard().PlacePiece({ 2, 3 }, pieceTwo, Game::Direction::EAST);
 
 		// Once we have added a second piece, still in the center row but facing east, all symmetries should be broken
-		auto [xSymmetryTwoPieces, ySymmetryTwoPieces] = GetBoardSymmetries(game);
+		auto [xSymmetryTwoPieces, ySymmetryTwoPieces] = GetBoardSymmetries(game.GetBoard());
 		EXPECT_FALSE(xSymmetryTwoPieces);
 		EXPECT_FALSE(ySymmetryTwoPieces);
 	}
@@ -47,7 +47,7 @@ namespace Alphalcazar::Strategy::MinMax {
 		Game::Game game = SetupGameForMinMaxTesting(Game::PlayerId::PLAYER_ONE, true, pieceSetups);
 
 		// A single piece on a corner breaks X and Y symmetries
-		auto [xSymmetry, ySymmetry] = GetBoardSymmetries(game);
+		auto [xSymmetry, ySymmetry] = GetBoardSymmetries(game.GetBoard());
 		EXPECT_FALSE(xSymmetry);
 		EXPECT_FALSE(ySymmetry);
 	}
@@ -60,7 +60,7 @@ namespace Alphalcazar::Strategy::MinMax {
 
 		// The perimeter piece is on the x-axis and faces west, meaning we should still have
 		// x-axis symmetry
-		auto [xSymmetry, ySymmetry] = GetBoardSymmetries(game);
+		auto [xSymmetry, ySymmetry] = GetBoardSymmetries(game.GetBoard());
 		EXPECT_TRUE(xSymmetry);
 		EXPECT_FALSE(ySymmetry);
 
@@ -69,7 +69,7 @@ namespace Alphalcazar::Strategy::MinMax {
 
 		// After placing a second piece (which on its own would still make the board have x-axis symmetry)
 		// we check that the combination of both pieces breaks both symmetries, each piece breaking 1 axis.
-		auto [xSymmetryTwoPieces, ySymmetryTwoPieces] = GetBoardSymmetries(game);
+		auto [xSymmetryTwoPieces, ySymmetryTwoPieces] = GetBoardSymmetries(game.GetBoard());
 		EXPECT_FALSE(xSymmetryTwoPieces);
 		EXPECT_FALSE(ySymmetryTwoPieces);
 	}
@@ -80,7 +80,7 @@ namespace Alphalcazar::Strategy::MinMax {
 		// On an empty board, all combinations of piece type and perimeter tile should create a legal move
 		EXPECT_EQ(legalMoves.size(), Game::c_PieceTypes * game.GetBoard().GetPerimeterTiles().size());
 
-		FilterSymmetricMovements(legalMoves, game);
+		FilterSymmetricMovements(legalMoves, game.GetBoard());
 
 		// After filtering symmetric movements on an empty board, we expect to only have 2 options for each piece
 		EXPECT_EQ(legalMoves.size(), Game::c_PieceTypes * 2);
@@ -93,7 +93,7 @@ namespace Alphalcazar::Strategy::MinMax {
 		Game::Game game = SetupGameForMinMaxTesting(Game::PlayerId::PLAYER_ONE, true, pieceSetups);
 
 		auto legalMoves = game.GetLegalMoves(Game::PlayerId::PLAYER_TWO);
-		FilterSymmetricMovements(legalMoves, game);
+		FilterSymmetricMovements(legalMoves, game.GetBoard());
 
 		/*
 		 * On a board with either x-axis or y-axis symmetry (but not both), the amount of available
@@ -115,12 +115,40 @@ namespace Alphalcazar::Strategy::MinMax {
 
 		auto legalMoves = game.GetLegalMoves(Game::PlayerId::PLAYER_ONE);
 		auto legalMovesSize = legalMoves.size();
-		FilterSymmetricMovements(legalMoves, game);
+		FilterSymmetricMovements(legalMoves, game.GetBoard());
 
 		// On this board, that has no symmetry, we expect all legal moves to be available after filtering
 		EXPECT_EQ(legalMoves.size(), legalMovesSize);
 		// The legal moves size should be equal to all perimeter tiles (they are all empty) times
 		// al pieces minus 2, as player 1 has already placed 2 pieces on the board
 		EXPECT_EQ(legalMoves.size(), (Game::c_PieceTypes - 2) * game.GetBoard().GetPerimeterTiles().size());
+	}
+
+	TEST(LegalMovements, SortLegalMovements) {
+		std::vector<PieceSetup> pieceSetups{
+			{ Game::PlayerId::PLAYER_ONE, 2, Game::Direction::SOUTH, { 1, 1 } }
+		};
+		Game::Game game = SetupGameForMinMaxTesting(Game::PlayerId::PLAYER_ONE, true, pieceSetups);
+
+		std::vector<Game::PlacementMove> legalMoves {
+			// This movement will result in the piece not entering the board
+			{ { 1, 0 }, 3 },
+			// This movement will enter in the center lane
+			{ { 2, 0 }, 3 },
+			// This movement will enter in a lateral lane
+			{ { 3, 0 }, 3 }
+		};
+		std::size_t sizeBeforeSorting = legalMoves.size();
+		SortLegalMovements(legalMoves, game.GetBoard());
+		std::size_t sizeAfterSorting = legalMoves.size();
+
+		EXPECT_EQ(sizeBeforeSorting, sizeAfterSorting);
+
+		// We expect the movement that causes the piece to enter in the center lane
+		// to be first, followed by the movement that causes the piece to enter the lateral lane,
+		// and the movement that causes the piece not to enter at all last
+		EXPECT_EQ(legalMoves[0].Coordinates.x, 2);
+		EXPECT_EQ(legalMoves[1].Coordinates.x, 3);
+		EXPECT_EQ(legalMoves[2].Coordinates.x, 1);
 	}
 }