--- a/cpp/src/strategies/random/src/random/RandomStrategy.cpp
+++ b/cpp/src/strategies/random/src/random/RandomStrategy.cpp
@@ -9,8 +9,8 @@ namespace Alphalcazar::Strategy::Random {
 		, mRandomEngine { mRandomDevice() }
 	{}
 
-	Game::PlacementMoveIndex RandomStrategy::Execute(Game::PlayerId, const std::vector<Game::PlacementMove>& legalMoves, const Game::Game&) {
-		std::uniform_int_distribution<std::mt19937::result_type> distribution { 0, static_cast<Game::PlacementMoveIndex>(legalMoves.size() - 1) };
-		return static_cast<Game::PlacementMoveIndex>(distribution(mRandomEngine));
+	Game::PlacementMove RandomStrategy::Execute(Game::PlayerId, const std::vector<Game::PlacementMove>& legalMoves, const Game::Game&) {
+		std::uniform_int_distribution<std::mt19937::result_type> distribution { 0, static_cast<std::mt19937::result_type>(legalMoves.size() - 1) };
+		return legalMoves[distribution(mRandomEngine)];
 	}
 }