Heuristic sorting of legal moves to try to get the most out of alpha-beta-pruning (it saves the most time if the options happen to be sorted from best to worst).

Did not affect performance very significantly (actually decreased it for depth 2), but it's a small win nonetheless, and it provides the base for exploring better sorting heuristics.

Before:
```
[2022-07-08 23:31:50.608] [info] First move at depth 1 took 0ms and calculated a score of 0
[2022-07-08 23:31:50.733] [info] First move at depth 2 took 124ms and calculated a score of -34
[2022-07-08 23:32:12.703] [info] First move at depth 3 took 21968ms and calculated a score of 0
```

After:
```
[2022-07-08 23:29:58.856] [info] First move at depth 1 took 0ms and calculated a score of 0
[2022-07-08 23:29:59.278] [info] First move at depth 2 took 421ms and calculated a score of -34
[2022-07-08 23:30:14.575] [info] First move at depth 3 took 15296ms and calculated a score of 0
```